filter blocks based on page permissions
